+++ app/api/workspace/teams/[...teamId]/route.ts
@@ -1,0 +1,36 @@
+import { AxiosErrorHandler } from "@/lib/errorHandler";
+import { workspaceService } from "@/services/workspace.service";
+import { NextRequest, NextResponse } from "next/server";
+
+
+
+export async function GET(req: NextRequest,
+    { params }: { params: Promise<{ teamId: string }> }
+) {
+    try {
+
+        const { teamId } = await params;
+        const { searchParams } = new URL(req.url);
+        const workspaceId = searchParams.get("workspaceId");
+
+        const res = await workspaceService.getTeam({ workspaceId, teamId });
+
+        return NextResponse.json({
+            success: res.data.success,
+            data: res.data.data,
+            message: res.data.message,
+        });
+    } catch (error) {
+        const err = AxiosErrorHandler(error);
+        return NextResponse.json(
+            {
+                success: false,
+                message: err.message,
+                errorCode: err.errorCode,
+            },
+            {
+                status: err.statusCode ?? 500,
+            }
+        );
+    }
+}
