+++ constants/routeHandler.ts
@@ -15,4 +15,7 @@ export const NEXT_API_ROUTES = {
   GET_WORKSPACE_MEMBERS: "/api/workspace/members",
   GET_WORKSPACE_INVITES: "/api/workspace/invites",
   GET_WORKSPACE_ROLES: "/api/workspace/roles",
+  GET_WORKSPACE_TEAMS: "/api/workspace/teams",
+  GET_WORKSPACE_TEAM: (teamId: string) => `/api/workspace/teams/${teamId}`,
+
 };
